requirements.txt
flask
openai
azure-search-documents
azure-core